Split core vs consistency
